Continue adjusting ui code structure
继续重构和查重部分的代码
